+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added priority and retro-compatibility for etc/autosubmitrc over etc/.autosubmitrc #2312
 - Updated Slurm wallclock truncation, allowing users to use e.g., 01:00:00 correctly #2059
 - Fixed exception trace in AutosubmitCritical and AutosubmitError not being displayed to the user #2109
+- Fix `Jobs.running` to case-insensitive #2916
 
 **Enhancements:**